3.9 增加调节录制音量功能

+++ Xpai/upLineViewController/VRCViewController/SettingView/ResolutionRatioView.m
@@ -74,7 +74,7 @@
 }
 
 -(UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath {
-    NSLog(@"count%d",_dataSource.count);
+    NSLog(@"count%lu",(unsigned long)_dataSource.count);
     resolutionRatioCell * cell = [resolutionRatioCell cellWithTableView:tableView];
     cell.contentLB.text = _dataSource[indexPath.row];
